refactor #3 : package 분리 및 navigation Base
