+++ app/google-services.json
@@ -1,0 +1,29 @@
+{
+  "project_info": {
+    "project_number": "552942230032",
+    "project_id": "baekyoung-i",
+    "storage_bucket": "baekyoung-i.appspot.com"
+  },
+  "client": [
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:552942230032:android:0cd2032c4dfe1a451ec895",
+        "android_client_info": {
+          "package_name": "com.tgyuu.baekyoung_i"
+        }
+      },
+      "oauth_client": [],
+      "api_key": [
+        {
+          "current_key": "AIzaSyBZcKzjFYf-6ttfKdH_WBfRSNq1Y1CJuxY"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": []
+        }
+      }
+    }
+  ],
+  "configuration_version": "1"
+}
