+++ app/src/main/java/com/tgyuu/baekyoung_i/auth/AuthScreen.kt
@@ -1,4 +1,4 @@
-package com.tgyuu.baekyoung_i
+package com.tgyuu.baekyoung_i.auth
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -21,11 +21,11 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.tgyuu.baekyoung_i.R
 import com.tgyuu.designsystem.theme.BaekyoungTheme
 import com.tgyuu.designsystem.theme.Blue00
 import com.tgyuu.designsystem.theme.Blue37
